Refactored the Specifics class and subclasses. These can now be changed/extended.
